Change price amount type to BigInt #302

+++ packages/api/schema/inputTypes.js
@@ -123,7 +123,7 @@ export default [
     }
 
     input UpdateProductCommercePricingInput {
-      amount: Int!
+      amount: BigInt!
       maxQuantity: Int
       isTaxable: Boolean
       isNetPrice: Boolean
